requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.6.1
python-dotenv==1.0.0
docker==6.1.3
pytest==7.4.3
pytest-asyncio>=0.23.0
websockets==12.0
python-multipart==0.0.6
redis==5.0.1
gitpython==3.1.40
speechrecognition==3.10.1
pyyaml==6.0.1
aiofiles==23.2.1
tenacity==8.2.3
loguru==0.7.2
pydantic==2.5.0
pydantic-settings==2.1.0
requests
gunicorn
httpx
PyAudio
prometheus-fastapi-instrumentator==6.1.0